# docs/api/user-bff.openapi.yaml
openapi: 3.0.3
info:
  title: User BFF API
  description: Backend for Frontend API para o aplicativo mobile
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api-stg.seudominio.com
    description: Staging

paths:
  /app/config:
    get:
      summary: Configuração do app (sem autenticação)
      operationId: getAppConfig
      tags:
        - Config
      responses:
        '200':
          description: Configuração do app
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiVersion:
                    type: string
                    example: "1.0.0"
                  minAppVersion:
                    type: string
                    example: "1.0.0"
                  features:
                    type: object
                    properties:
                      devMode:
                        type: boolean
                        example: true

  /me:
    get:
      summary: Informações do usuário autenticado
      operationId: getMe
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /wallets/summary:
    get:
      summary: Resumo da carteira (saldo)
      operationId: getWalletSummary
      tags:
        - Wallet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resumo da carteira
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /transactions:
    get:
      summary: Lista de transações
      operationId: getTransactions
      tags:
        - Transactions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Número máximo de transações
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista de transações
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  /transactions/{id}:
    get:
      summary: Detalhes de uma transação
      operationId: getTransactionById
      tags:
        - Transactions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes da transação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /dev/transactions/generate:
    post:
      summary: Gerar transações de teste (apenas dev)
      operationId: generateTransactions
      tags:
        - Dev
      security:
        - bearerAuth: []
      parameters:
        - name: count
          in: query
          description: Número de transações a gerar
          required: false
          schema:
            type: integer
            default: 1000
            minimum: 1
            maximum: 10000
      responses:
        '200':
          description: Transações geradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated:
                    type: integer
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Endpoint disponível apenas em ambiente dev

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        name:
          type: string

    WalletSummary:
      type: object
      properties:
        balance:
          type: number
          format: double
          example: 1250.50
        currency:
          type: string
          example: "BRL"
        lastUpdated:
          type: string
          format: date-time

    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [DEBIT, CREDIT, TOPUP]
        amount:
          type: number
          format: double
        merchant:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]

    TransactionDetail:
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            reference:
              type: string
            metadata:
              type: object

  responses:
    Unauthorized:
      description: Não autenticado ou token inválido
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"

    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Not found"